search_changes

+++ services/ProductService.js
@@ -10,16 +10,16 @@ class ProductService {
         try {
             let sql = `SELECT * FROM products t1 INNER JOIN ${category_name} t2 ON t1.product_id = t2.product_id  
                             WHERE t1.category_name = "${category_name}" && t1.product_id = ${product_id}`
-            let result = await db.promise().query(sql)
+            let result = await db.query(sql)
             if (!result?.[0]?.[0]) {
                 throw ApiError.BadRequest('Unknown category/product specified')
             }
             sql = `SELECT img, master FROM imgs WHERE product_id = ${product_id}`
-            let imgs = await db.promise().query(sql)
+            let imgs = await db.query(sql)
             sql = `SELECT ref_num, brand, product_id, ref_id FROM refs WHERE product_id = ${product_id}`
-            let refs = await db.promise().query(sql)
+            let refs = await db.query(sql)
             sql = `SELECT oem, model, product_id, oem_id FROM oems WHERE product_id = ${product_id}`
-            let oems = await db.promise().query(sql)
+            let oems = await db.query(sql)
 
             return {
                 ...result[0][0],
@@ -35,7 +35,7 @@ class ProductService {
     async getAllProducts() {
         try {
             let sql = `SELECT p.*, i.img FROM products p INNER JOIN imgs i ON p.product_id = i.product_id WHERE i.master = 1`
-            let result = await db.promise().query(sql)
+            let result = await db.query(sql)
             return result[0]
         } catch (err) {
             throw ApiError.ServerError(err.message)
@@ -50,7 +50,7 @@ class ProductService {
                             WHERE t1.category_name = "${category_name}" 
                             AND t1.product_id = i.product_id
                             AND i.master = 1`
-            const result = await db.promise().query(sql)
+            const result = await db.query(sql)
             return result?.[0]
         } catch (err) {
             throw ApiError.BadRequest(err.message)
@@ -73,19 +73,20 @@ class ProductService {
             price_copy,
             price_original,
             discount,
+            top_selling,
             ...rest
         } = data
 
         try {
             const fileNames = []
             let sql = `SELECT * FROM categories WHERE category_id = ${category_id} && name_ = "${category_name}"`
-            let result = await db.promise().query(sql)
+            let result = await db.query(sql)
             if (!result?.[0]?.[0]) {
                 throw ApiError.BadRequest('Unknown category specified')
             }
 
             sql = `INSERT INTO products SET ?`
-            result = await db.promise().query(sql, {
+            result = await db.query(sql, {
                 brand,
                 model,
                 detail_number,
@@ -95,23 +96,29 @@ class ProductService {
                 count_copy,
                 price_original,
                 price_copy,
+                top_selling
             })
             let insertId = result[0].insertId
 
-            const refValues = refs.map((ref) => [
-                ref.ref_num,
-                ref.brand,
-                insertId,
-            ])
-            sql = `INSERT INTO refs (ref_num, brand, product_id) VALUES ?`
-            await db.promise().query(sql, [refValues])
+            if(refs.length){
+                const refValues = refs.map((ref) => [
+                    ref.ref_num,
+                    ref.brand,
+                    insertId,
+                ])
+                sql = `INSERT INTO refs (ref_num, brand, product_id) VALUES ?`
+                await db.query(sql, [refValues])    
+            }
+            
+            if(oems.length){
+                const oemValues = oems.map((item) => {
+                    return [item.oem, item.model, insertId]
+                })
 
-            const oemValues = oems.map((item) => {
-                return [item.oem, item.model, insertId]
-            })
+                sql = `INSERT INTO oems (oem, model, product_id) VALUES ?`
+                await db.query(sql, [oemValues])                
+            }
 
-            sql = `INSERT INTO oems (oem, model, product_id) VALUES ?`
-            await db.promise().query(sql, [oemValues])
 
             imgs.forEach((img) => {
                 const fileName = `${uuid.v4()}.${img.mimetype.split('/')[1]}`
@@ -125,10 +132,10 @@ class ProductService {
             ])
 
             sql = `INSERT INTO imgs (img, master, product_id) VALUES ?`
-            await db.promise().query(sql, [imgValues])
+            await db.query(sql, [imgValues])
 
             sql = `INSERT INTO ${category_name} SET ?`
-            await db.promise().query(sql, { ...rest, product_id: insertId })
+            await db.query(sql, { ...rest, product_id: insertId })
             return true
         } catch (err) {
             throw ApiError.BadRequest(err.message)
@@ -144,7 +151,7 @@ class ProductService {
                 ...restData
             } = data
             let sql = `SELECT * FROM products WHERE product_id = ${prId} && category_name = "${catName}"`
-            let result = await db.promise().query(sql)
+            let result = await db.query(sql)
             if (!result[0][0]) {
                 throw ApiError.BadRequest('Unknown category/product specified')
             }
@@ -165,6 +172,7 @@ class ProductService {
                 price_original,
                 price_copy,
                 detail_number,
+                top_selling,
                 ...rest
             } = restData
 
@@ -174,21 +182,26 @@ class ProductService {
 
             deleted_imgs = JSON.parse(deleted_imgs)
 
-            const refValues = refs.map((ref) => [ref.ref_num, ref.brand, prId])
-            if (refValues.length) {
-                sql = `DELETE FROM refs WHERE product_id = ${prId}`
-                await db.promise().query(sql)
-                sql = `INSERT INTO refs (ref_num, brand, product_id) VALUES ?`
-                await db.promise().query(sql, [refValues])
+            if(refs.length){
+                const refValues = refs.map((ref) => [ref.ref_num, ref.brand, prId])
+                if (refValues.length) {
+                    sql = `DELETE FROM refs WHERE product_id = ${prId}`
+                    await db.query(sql)
+                    sql = `INSERT INTO refs (ref_num, brand, product_id) VALUES ?`
+                    await db.query(sql, [refValues])
+                }
             }
-
-            const oemValues = oems.map((item) => {
-                return [item.oem, item.model, prId]
-            })
-            sql = `DELETE FROM oems WHERE product_id = ${prId}`
-            await db.promise().query(sql)
-            sql = `INSERT INTO oems (oem, model, product_id) VALUES ?`
-            await db.promise().query(sql, [oemValues])
+            
+            if(oems.length){
+                const oemValues = oems.map((item) => {
+                    return [item.oem, item.model, prId]
+                })
+                sql = `DELETE FROM oems WHERE product_id = ${prId}`
+                await db.query(sql)
+                sql = `INSERT INTO oems (oem, model, product_id) VALUES ?`
+                await db.query(sql, [oemValues])    
+            }
+            
 
             let propsToUpdate = ``
             let restPropsToUpdate = ``
@@ -216,11 +229,11 @@ class ProductService {
             propsToUpdate = propsToUpdate.slice(0, -2)
             if (propsToUpdate) {
                 sql = `UPDATE products SET ${propsToUpdate} WHERE product_id = ${prId}`
-                await db.promise().query(sql)
+                await db.query(sql)
             }
             if (restPropsToUpdate) {
                 sql = `UPDATE ${catName} SET ${restPropsToUpdate} WHERE product_id = ${prId}`
-                await db.promise().query(sql)
+                await db.query(sql)
             }
             if (deleted_imgs.length) {
                 sql = `DELETE FROM imgs WHERE product_id = ${prId} AND img IN(`
@@ -238,7 +251,7 @@ class ProductService {
                 })
                 sql = sql.slice(0, -2)
                 sql += ')'
-                await db.promise().query(sql)
+                await db.query(sql)
             }
 
             if (imgs.length) {
@@ -254,11 +267,11 @@ class ProductService {
 
                 sql = `INSERT INTO imgs (img, product_id) VALUES ?`
 
-                await db.promise().query(sql, [imgValues])
+                await db.query(sql, [imgValues])
             }
             sql = `SELECT * FROM imgs WHERE product_id = ${prId}`
             let updateMsterImgSql = ''
-            const productImgs = await db.promise().query(sql)
+            const productImgs = await db.query(sql)
             if (productImgs?.[0].length) {
                 productImgs?.[0].forEach((imgItem, index) => {
                     if (index == img_master) {
@@ -266,7 +279,7 @@ class ProductService {
                     }
                 })
                 if (updateMsterImgSql) {
-                    await db.promise().query(updateMsterImgSql)
+                    await db.query(updateMsterImgSql)
                 }
             }
         } catch (err) {
@@ -279,7 +292,7 @@ class ProductService {
         const { category_name, product_id } = data
         try {
             let sql = `SELECT * FROM imgs WHERE product_id = "${product_id}"`
-            const res = await db.promise().query(sql)
+            const res = await db.query(sql)
             if (res?.[0].length) {
                 res?.[0].forEach((productItem) => {
                     const pathToImg = path.resolve(
@@ -294,7 +307,7 @@ class ProductService {
                 })
             }
             sql = `DELETE FROM products WHERE category_name = "${category_name}" && product_id = ${product_id}`
-            await db.promise().query(sql)
+            await db.query(sql)
             return true
         } catch (err) {
             throw ApiError.BadRequest(err.message)
@@ -311,7 +324,7 @@ class ProductService {
             refsToRemove += ")";
 
             let sql = `DELETE FROM refs WHERE product_id = ${product_id} AND ref_num IN ${refsToRemove}`
-            await db.promise().query(sql)
+            await db.query(sql)
         } catch (err) {
             throw ApiError.ServerError(err.message)
         }
@@ -327,7 +340,7 @@ class ProductService {
             oemsToRemove += ")";
 
             let sql = `DELETE FROM oems WHERE product_id = ${product_id} AND oem IN ${oemsToRemove}`
-            await db.promise().query(sql)
+            await db.query(sql)
         } catch (err) {
             throw ApiError.ServerError(err.message)
         }
@@ -336,7 +349,11 @@ class ProductService {
     async deleteImg({ product_id, img }) {
         try {
             let sql = `DELETE FROM imgs WHERE product_id = ${product_id} AND img = "${img}"`
-            await db.promise().query(sql)
+            await db.query(sql)
+                        if(nextImg){
+                sql = `UPDATE imgs SET master = 1 WHERE img = "${nextImg}"`
+                await db.query(sql)
+            }
             const imgPath = path.resolve(__dirname, '..', 'static', img)
             if (fs.existsSync(imgPath)) {
                 fs.unlinkSync(imgPath)
@@ -349,7 +366,7 @@ class ProductService {
     async getSparkPlugReferences(sparkPlug_id) {
         try {
             let sql = `SELECT * FROM products t1 INNER JOIN spark_plugs t2 ON t1.product_id = t2.product_id WHERE t1.product_id = ${sparkPlug_id} && t1.category_name = "spark_plugs"`
-            let result = await db.promise().query(sql)
+            let result = await db.query(sql)
             if (!result?.[0]?.[0]) {
                 throw ApiError.BadRequest('Unknown spark plug specified')
             }
@@ -367,7 +384,7 @@ class ProductService {
                             && thread_size = ${thread_size} && thread_length = ${thread_length} 
                             && electrodes_number = "${electrodes_number}" && t2.product_id != ${sparkPlug_id}
                     `
-            result = await db.promise().query(sql)
+            result = await db.query(sql)
             return result[0]
         } catch (err) {
             throw ApiError.BadRequest(err.message)
@@ -375,58 +392,53 @@ class ProductService {
     }
 
     async searchProduct(searchData) {
-        try {
-            //search by ref_num
-            let sql = `SELECT * FROM products WHERE detail_number = "${searchData}" OR detail_number = "${searchData}"`
-            let result = await db.promise().query(sql)
-            if (result[0].length) {
-                const { category_name } = result[0][0]
-                sql = `SELECT * FROM products p 
-                            INNER JOIN ${category_name} t2 ON p.product_id = t2.product_id
-                            INNER JOIN imgs i ON p.product_id = i.product_id
-                            WHERE p.detail_number = "${searchData}" AND i.master = 1`
-                let product = await db.promise().query(sql)
-                return product[0]
+        try{
+            let sql
+            let resultsToReturn = []
+            const searchDataWithoutChars = searchData.split(' ').join('').split('-').join('')
+
+            sql = `SELECT DISTINCT p.*, i.img FROM products p INNER JOIN imgs i ON i.product_id = p.product_id 
+                WHERE (REPLACE(REPLACE(p.model, '-', ''), ' ', '') = '${searchDataWithoutChars}' 
+                OR REPLACE(REPLACE(p.detail_number, '-', ''), ' ', '') = '${searchDataWithoutChars}')
+                AND i.master = '1'    
+               `
+
+            let trialResult = await db.query(sql)
+            resultsToReturn = trialResult[0]
+
+            if(resultsToReturn.length){
+                return resultsToReturn
             }
 
-            //search by detail_number
-            sql = `SELECT * FROM products p INNER JOIN refs r ON p.product_id = r.product_id WHERE r.ref_num = "${searchData}" AND r.master = 1`
-            result = await db.promise().query(sql)
+            sql = `SELECT DISTINCT p.*, i.img FROM oems o INNER JOIN products p ON o.product_id = p.product_id
+                    INNER JOIN imgs i ON i.product_id = o.product_id
+                    WHERE REPLACE(REPLACE(o.oem, '-', ''), ' ', '') = '${searchDataWithoutChars}' 
+                    AND i.master = '1'    
+               `
 
-            if (result[0].length) {
-                let dataToReturn = result[0].map(async (item) => {
-                    const { category_name } = item
-                    sql = `SELECT * FROM products t1 INNER JOIN ${category_name} t2 ON t1.product_id = t2.product_id
-                            WHERE t1.product_id = "${item.product_id}"`
-                    const res = await db.promise().query(sql)
-                    return res[0][0]
-                })
-                dataToReturn = await Promise.all(dataToReturn)
-                return dataToReturn
-            } else {
-                return []
+            trialResult = await db.query(sql)
+            resultsToReturn = trialResult[0]
+
+            if(resultsToReturn.length){
+                return resultsToReturn
             }
 
-            //search by oem
-            sql = `SELECT * FROM products t1 INNER JOIN oems t2 ON t1.product_id = t2.product_id WHERE t2.oem = ${searchData}`
-            result = await db.promise().query(sql)
-            if (result[0].length) {
-                let promises = result[0].map(async (item) => {
-                    const { category_name } = item
-                    sql = `SELECT * FROM products t1 INNER JOIN ${category_name} t2 ON t1.product_id = t2.product_id
-                            WHERE t1.product_id = t2.product_id`
-                    const res = await db.promise().query(sql)
-                    return res[0][0]
-                })
-                const allPromises = await Promise.all(promises)
-                const dataToReturn = {}
-                allPromises.forEach((item) => {
-                    dataToReturn[item.product_id] = item
-                })
-                return Object.values(dataToReturn)
+            sql = `SELECT DISTINCT p.*, i.img FROM refs r INNER JOIN products p ON r.product_id = p.product_id
+                    INNER JOIN imgs i ON i.product_id = r.product_id
+                    WHERE REPLACE(REPLACE(r.ref_num, '-', ''), ' ', '') = '${searchDataWithoutChars}' 
+                    AND i.master = '1'    
+               `
+
+            trialResult = await db.query(sql)
+            resultsToReturn = trialResult[0]
+
+            if(resultsToReturn.length){
+                return resultsToReturn
             }
-        } catch (err) {
-            throw ApiError.BadRequest(err.message)
+
+            return 'Unknown data'
+        }catch(err){
+            console.log({err})
         }
     }
 }
